Dev ilya

+++ src/project/project.controller.ts
@@ -9,6 +9,7 @@ import {
   UseGuards,
   Request,
   NotFoundException,
+  BadRequestException,
 } from '@nestjs/common';
 import { ProjectService } from './project.service';
 import { CreateProjectDto } from './dto/create-project.dto';
@@ -22,6 +23,7 @@ import {
   ApiTags,
 } from '@nestjs/swagger';
 import { ProjectEntity } from './entities/project.entity';
+import * as uuid from 'uuid';
 
 @Controller('projects')
 @ApiTags('projects')
@@ -54,6 +56,9 @@ export class ProjectController {
   @ApiOkResponse({ type: ProjectEntity })
   @ApiParam({ name: 'project', required: true })
   async findOne(@Request() req, @Param() params) {
+    if (!uuid.validate(params.project)) {
+      throw new BadRequestException();
+    }
     return new ProjectEntity(
       await this.projectService.findOne(req.user.email, params.project),
     );
@@ -69,6 +74,9 @@ export class ProjectController {
     @Param() params,
     @Body() updateProjectDto: UpdateProjectDto,
   ) {
+    if (!uuid.validate(params.project)) {
+      throw new BadRequestException();
+    }
     return await this.projectService.update(
       req.user.email,
       params.project,
@@ -82,6 +90,9 @@ export class ProjectController {
   @ApiOkResponse()
   @ApiParam({ name: 'project', required: true })
   async remove(@Request() req, @Param() params) {
+    if (!uuid.validate(params.project)) {
+      throw new BadRequestException();
+    }
     return await this.projectService.remove(req.user.email, params.project);
   }
 }
